Modularize plugin traits and mock plugin
