+++ client-base/src/plugin/has_turn.rs
@@ -1,0 +1,5 @@
+/// Indicates that the value has a turn.
+pub trait HasTurn {
+    /// Fetches the turn.
+    fn turn(&self) -> u32;
+}
